// src/api/coreTypes.ts
export interface RoomCreateType {
  name: string;
  type: number  | null;
  capacity?: number | null;
  branchId: number;
}

export interface RoomEditType {
  id: number;
  name?: string | null;
  type?: number | null;
  capacity?: number | null;
  branchId: number;
}

// ============== Category Types ==============
export interface CategoryGetByIdType {
  id: number;
  name: string;
  description?: string | null;
}

export interface CategoryCreateType {
  name: string;
  description?: string | null;
}

export interface CategoryEditType {
  id: number;
  name?: string | null;
  description?: string | null;
}